Better error handling in launchers.
Launchers generally worked by redirecting all non-job output to a launcher
log and passing job stout and stderr through. The executor, however, did
not read launcher logs (and by default even redirected it to /dev/null).
This mean that any errors in the launcher would have to potential to go
unnoticed.

This commit updates the launcher infrastructure as follows:
- adds two generic launcher methods that allow launchers to distinguish
between a launcher an job failure when a launcher exits with non-zero exit
code as well as a method allowing the extraction of the relevant error
message.
- implements this in the script based launchers by printing a magic line
at the end which signals successful completion of the launcher
- handles the above in the local executor
- changes the job redirection to happen in the launcher scripts; this
is needed so that no additional signaling file is needed for the launcher
errors.