Fix a race condition.
It is possible for a job to be registered after the status command is
started but before the copy of `self._jobs` is made. This could result
in the code interpreting the job as missing and consequently completed,
which would be wrong.

The fix makes a copy of `self._jobs` before the status command is invoked
as well as using the copy to generate the job ids to query the status
for.