Fix installation bug
When pip-installing the package from a clean repository, the
launcher scripts were not moved into the install directory.
This likely has to do with the timing of the 'install' class's
'run' method (i.e. that method probably runs after all files are
expected to be built). If you install twice, the launcher scripts
will be present from the first install, and so everything will
work as expected---this is why a clean repository is necessary
to reproduce the issue.

Resolve the problem by moving the script-building into the
build_py command.