Revamped the plugin loading mechanism.
Plugins are now loaded indirectly, using a descriptor class. This was
needed because one could not find details about, say, an executor (such
as name and version) if the executor file failed to import. However, a
minimal descriptor is virtually guaranteed to load, providing information
about the implementation that does not depend on that implementation being
able to load/import.

Also updated the way loading works. Previously, problems could occur when
implementations existed in multiple locations, because package traversal
in python is weird: despite correctly iterating over packages from
multiple locations, python only really references the first location
loaded and only allows imports from that location. The current
implementation works according to the following rules:
- descriptors must be in the 'psij.executors' package, and the package can
be spread across multiple locations
- implementations must exist in the same location as the descriptor; for
example, a descriptor in the system dist-packages can only be used to load
an implementation from the system dist-packages, not from a user directory.
- if implementations exist in both the same location and another location
as a descriptor, only the one in the same location will be used
- here, "location" means one directory in PYTHONPATH and/or sys.path.

And last, you can now run "python -m psij plugins" to get a detailed list
of what is loaded and from where.