Rewritten the descriptor loading code.
Python does not work well when the same package is split in multiple
places. The way pip deals with this is by actually merging packages
within the same directory (even when belonging to different products).
However, this is not a reliable option when one has to split a package
in both system directories as well as user ones.

The official solution is "namespace packages". Simply put, one omits
`__init__` from such packages. But, if one were to have `psij` as a
namespace package, there could be no initialization code for the package,
nor re-exports of deep modules into the `psij` namespace.

The solution adopted by this commit (which changes the previous solution
in this branch) is to move descriptors into a namespace package
independent of the `psij` package, namely `psij-descriptors`. This is
a purposefully illegal name, since descriptors should not be imported
directly. The exact loading mechanism used does not require the
`psij-descriptors` package to be a namespace package (i.e., it could have
an `__init__.py` file). However, this terribly upsets mypy, who then
refuses to check anything else, including inside properly named packages.

The implementations of executors/launchers must still exist in proper
packages. For any externally contributed implementations, the root package
name must be unique to avoid clashes with other implementations that may
be installed under a different path.