Moved PYTHONPATH setup from psij-ci-run to ci_runner.
One of the issues was that .packages was added to PYTHONPATH in psij-ci-run
after which a new at command was issued invoking psij-ci-run. This would,
in time, grow PYTHONPATH to the limit, after which any attempt to add to
it will result in error, and error which subtly results in failures with
loading extra executors in tests/.

Also use Path.resolve() instead of Path.absolute(), since the latter does
not appear to be documented.

Note: psij-ci-run is updated automatically on test sites. It needs to be
done manually. On sites that use "at", an extra step of clearing PYTHONPATH
should be performed (in, e.g., psij-ci-load).