Revert "tests: added tests for getting-started section of online docs"
Reverting this because it is freezing the CI on multiple machines.

One of the tests uses a shared global state without proper concurrency
controls, so it deadlocks on occasion, which leads to the test never
completing, which prevents the CI from running anything else. While the
CI should be resilient against such problems, the CI timing out a test
should be an issue of last resort.

This revert also removes a small number of other tests. The initial
idea that prompted this commit was to automatically test code that
is used in the documentation. However, this commit does not do so.
Instead, it adds some tests that are similar but not the same as what
is in the documentation, and are not automatically updated when
documentation tests are updated. Additionally, they do not test
anything the isn't already tested by the existing tests in the suite.

This reverts commit 99f4fe244305b994189802ea43d672264d7e70fe.