Fix text squeezing on large displays.
On large displays (with width >>14in), the padding kicks in. It is meant
to shift the text towards the center while preserving the width of the
main text (through the max-width property). However, the padding is part
of the width, so it ends up reducing the usable text width.

We could switch to controlling the margin, but then the thin left border
also moves, which is unpleasant. I guess we could do that AND have the
sidebar draw the border, but that's two steps.

So anyway, this adjusts the max-width to take the padding into account.