Added a custom generator for the API doc, since sphinx autodoc lacks the flexibility to produce nice results and manually doing the full api doc is tedious and error prone.
This generator takes a file that represents a toc tree and outputs an api
doc that results in (roughly) the same toc tree, but with proper class,
module, and function docs. It also procudes a list of executors and
launchers, as well as their respective class docs, by querying the
appropriate psij executor/launcher list functions.

Last, it produces an index of all classes and functions.